+++ packages/pluggableWidgets/barcode-scanner-web/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
--   We fixed issue with camera, when it keep capturing video after widget is closed.
+-   We fixed an issue where the camera kept capturing after the widget was closed.
 
 ## [2.2.1] - 2022-07-27